+++ src/Pages/HomePage/Components/BodyMapPreview/BodyMapPreviewStyle.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  container: {
+    width: "58%",
+    maxWidth: 210,
+    aspectRatio: 503 / 1294,
+    alignSelf: "center",
+    position: "relative",
+  },
+});